Deeperbeige / Dec 30, 2014 / I was dismayed to find that the new 4.6 uGUI Text component supports line spacing, but not letter spacing. Especially since the later versions of NGUI had this option. I figured that when the UI source came out, I could probably add it pretty easily. Nope - the text layout engine isn't part of the open source stuff, and you can't get at the code or a suitable API for it.
My best workaround was to simply modify existing fonts and add extra space to every character. This is as difficult to manage as it sounds, and impractical for a recent design I'm working to which has lots of differently spaced headings. I would need a different font for each different spacing.

Then by chance I looked at the source for the Shadow effect, which uses a BaseVertexEffect to modify the list of vertices that make the geometry of the object being drawn. An hour or two of hacking, and I've got a component that spaces text out via much the same technique (although it doesn't generate any extra verts). Add the component to a UI textfield, adjust the spacing value, job done.

I thought I should share the effect here, as I've seen a few other threads asking the same basic question.

While this effect works in the editor, plays well on mobile and integrates well with the new UI system, it does have its limits. It doesn't work well with rich text, or automatically wrapped text. Full details in the comments at the top of the code, but don't expect miracles. If you can improve some part of it, let me know and I'll update it for everyone else.

Download the effect here:
http://deeperbeige.com/files/unity/LetterSpacing.zip