huulong / Feb 24, 2016 / I have not checked InputField so the caret problem is most probably still there, but I merged ArshakKroyan's Rich Text support with The Paper's Unity 5.3+ support. Additionally I made the following modifications: Support Rich Text on multiple lines Now you can use Rich Text tags on the 2nd, 3rd lines, etc. and also span a single tag over multiple lines. Just changed the
Code (CSharp):

    charIdx = currentMatchedTag.Length + currentMatchedTag.Index - 1;
    glyphIdx = currentMatchedTag.Length + currentMatchedTag.Index - 1;
    charPositionIndex--;

to the corresponding relative operations (I also renamed charPositionIndex to actualCharIndex to show it's the index of a non-tag character):

Code (CSharp):

    // skip matched RichText tag
    charIdx += currentMatchedTag.Length - 1;
    actualCharIndex--;
    glyphIdx += currentMatchedTag.Length;
    continue;

(decrements are just here to cancel natural for loop increments; indices may become -1 if the text starts with a tag, but continue ensures the loop body is not executed anyway)

This way,
glyphIdx will not be reset on each line iteration and new lines will be correctly offset instead of characters being offset from the 1st line repeatedly. Note that I also immediately continue if we are dealing with a RichText tag, since offsetting an invisible character is not useful.

    Later versions of Unity 5 will use the Unity 5.2 and 5.3 branches of the script

UPDATE: use UNITY_X_Y_OR_NEWER macro instead of "#if UNITY_5 && !UNITY_5_0 && !UNITY_5_1"

I just used a complementary test

Code (CSharp):

    #if UNITY_5_2_OR_NEWER

instead of

Code (CSharp):

    #if UNITY_5_2 || UNITY_5_3

Just revert these parts if you prefer stating explicit versions.

I'm not sure why The Paper uses
Code (CSharp):

    if (idx4 > verts.Count - 1) return;

for the truncated text test and others use idx6. Since vertices come as a bundle I guess it doesn't matter.

It starts getting neat I think (even if the original author considers this as a terrible hack), any plan to release it on some Wiki, public repository (such as UnityToolbag) or something? Just to avoid copy-pasting the new version of the code each time. By the way I also attached the script if you prefer downloading it.

UPDATE: please replace "#if UNITY_5 && !UNITY_5_0 && !UNITY_5_1" with "if UNITY_5_2_OR_NEWER" in the script if you download it